crewai/skeleton/requirements.txt
crewai>=0.11.2
langchain>=0.1.16
langchain_community>=0.0.38
requests>=2.31.0
pydantic>=2.10.6
python-dotenv>=1.0.1
pyyaml>=6.0.1
duckduckgo-search>=3.9.4
fastapi>=0.110.0
uvicorn>=0.29.0
markdown
tabulate
